Avoid requirement for root privileges.
Root privileges are not required anymore because cryptsetup is called in
dry-run mode which does not actually change device mapper configuration.

This should work under non-root user as long as the user can access the
device. You can e.g. copy LUKS header to a file and run the cracker on
this file.
$ dd if=/dev/source of=/tmp/luks-header bs=1M count=10